sync codes with master

+++ process/recognition/GetCos.cpp
@@ -2,7 +2,7 @@
 
 double GetVectorCos(const std::vector<double>& v1, double* v2)
 {
-    int v1_size = v1.size();
+    unsigned long long v1_size = v1.size();
 //        if (v1_size != v2.size())
 //        {
 //            return 0;
@@ -11,7 +11,7 @@ double GetVectorCos(const std::vector<double>& v1, double* v2)
     double numerator = 0;
     double denominator_l = 0, denominator_r = 0;
 
-    for (int i = 0; i < v1_size; i++)
+    for (unsigned long long i = 0; i < v1_size; i++)
     {
         numerator += v1.at(i) * v2[i];
         denominator_l += v1.at(i) * v1.at(i);
